feat: bubble click event up through target tree
As commented in https://github.com/gregberge/react-teleporter/issues/7, there
might be use-cases where click bubbling is needed up through the location where
the portal is rendered (aka the Target) not in the React/component tree that
calls the portal.

Consider, for example, a slide-in sidebar menu that holds contextual actions
that need to be aware of when the user clicks on them in order to hide itself
without additional context or even callbacks.