feat: remove the box generated by source wrapper
In the previous commit the content of the Source was wrapped in a div in order
to be able to intercept the events bubbling from its content. To minimize its
side effects and avoid breaking the layout, the "display: contents;" CSS rule
is shipped in an inline style. To know more about it, please refer to the links
below.

  * https://drafts.csswg.org/css-display/#valdef-display-contents
  * https://developer.mozilla.org/en-US/docs/Web/CSS/display#display_contents
  * https://caniuse.com/css-display-contents
  * https://bitsofco.de/how-display-contents-works/
  * https://css-tricks.com/get-ready-for-display-contents/