feat: allows choosing events to be forwarded
Instead of always forwarding the click event, a `forwardToTarget` option has
been added to the `createTeleporter` method. This option adds more flexibility
in which events should be intercepted and sent to the Target. None by default,
to keep the library behavior almost as it was (except for the unboxed Source
wrapper).